+++ edge/mem_test.go
@@ -1,0 +1,22 @@
+// Copyright 2023 - MinIO, Inc. All rights reserved.
+// Use of this source code is governed by the AGPLv3
+// license that can be found in the LICENSE file.
+
+package edge_test
+
+import (
+	"testing"
+
+	"github.com/minio/kes/internal/keystore/mem"
+)
+
+func TestInMem(t *testing.T) {
+	ctx, cancel := testingContext(t)
+	defer cancel()
+
+	store := &mem.Store{}
+	t.Run("Create", func(t *testing.T) { testCreate(ctx, store, t) })
+	t.Run("Set", func(t *testing.T) { testSet(ctx, store, t) })
+	t.Run("Get", func(t *testing.T) { testGet(ctx, store, t) })
+	t.Run("Status", func(t *testing.T) { testStatus(ctx, store, t) })
+}
